# back/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/products": {
      "get": {
        "summary": "Get Admin Products",
        "operationId": "get_admin_products_api_admin_products_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProductInDB"
                  },
                  "type": "array",
                  "title": "Response Get Admin Products Api Admin Products Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Admin Product",
        "operationId": "create_admin_product_api_admin_products_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductInDB"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/products/{product_id}": {
      "get": {
        "summary": "Get Admin Product",
        "operationId": "get_admin_product_api_admin_products__product_id__get",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductInDB"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Admin Product",
        "operationId": "delete_admin_product_api_admin_products__product_id__delete",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Admin Product",
        "operationId": "update_admin_product_api_admin_products__product_id__patch",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Product Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductInDB"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/cart/merge": {
      "post": {
        "summary": "Merge Cart",
        "operationId": "merge_cart_api_cart_merge_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CartMergeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CartItemInDB"
                  },
                  "type": "array",
                  "title": "Response Merge Cart Api Cart Merge Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/categories": {
      "get": {
        "summary": "Get Categories",
        "operationId": "get_categories_api_categories_get",
        "parameters": [
          {
            "name": "slug",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryInDB"
                  },
                  "title": "Response Get Categories Api Categories Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/orders": {
      "get": {
        "summary": "Get User Orders",
        "description": "Получить заказы текущего пользователя",
        "operationId": "get_user_orders_api_orders_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OrderInDB"
                  },
                  "type": "array",
                  "title": "Response Get User Orders Api Orders Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Order",
        "operationId": "create_order_api_orders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderInDB"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Order",
        "operationId": "delete_order_api_orders_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderDelete"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CartItemInDB": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "productId": {
            "type": "string",
            "format": "uuid",
            "title": "Productid"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "title": "Userid"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "priceSnapshot": {
            "type": "number",
            "title": "Pricesnapshot"
          }
        },
        "type": "object",
        "required": [
          "id",
          "productId",
          "userId",
          "quantity",
          "priceSnapshot"
        ],
        "title": "CartItemInDB"
      },
...
      "CartMergeRequest": {
        "properties": {
          "localCart": {
            "items": {
              "$ref": "#/components/schemas/LocalCartItem"
            },
            "type": "array",
            "title": "Localcart"
          }
        },
        "type": "object",
        "required": [
          "localCart"
        ],
        "title": "CartMergeRequest"
      },
...
      "CategoryInDB": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "imageUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imageurl"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "slug",
          "id"
        ],
        "title": "CategoryInDB"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LocalCartItem": {
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid",
            "title": "Productid"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "priceSnapshot": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Pricesnapshot"
          }
        },
        "type": "object",
        "required": [
          "productId",
          "quantity",
          "priceSnapshot"
        ],
        "title": "LocalCartItem"
      },
      "OrderCreate": {
        "properties": {
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "postal_code": {
            "type": "string",
            "title": "Postal Code"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "total_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Total Amount"
          },
          "order_items": {
            "items": {
              "$ref": "#/components/schemas/OrderItemCreate"
            },
            "type": "array",
            "title": "Order Items"
          }
        },
        "type": "object",
        "required": [
          "full_name",
          "email",
          "address",
          "city",
          "postal_code",
          "phone",
          "total_amount",
          "order_items"
        ],
        "title": "OrderCreate"
      },
      "OrderDelete": {
        "properties": {
          "orderId": {
            "type": "string",
            "format": "uuid",
            "title": "Orderid"
          }
        },
        "type": "object",
        "required": [
          "orderId"
        ],
        "title": "OrderDelete"
      },
      "OrderInDB": {
        "properties": {
          "fullName": {
            "type": "string",
            "title": "Fullname"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "postalCode": {
            "type": "string",
            "title": "Postalcode"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "totalAmount": {
            "type": "string",
            "title": "Totalamount"
          },
          "orderItems": {
            "items": {
              "$ref": "#/components/schemas/OrderItemInDB"
            },
            "type": "array",
            "title": "Orderitems"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "title": "Userid"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "fullName",
          "email",
          "address",
          "city",
          "postalCode",
          "phone",
          "totalAmount",
          "id",
          "userId",
          "status",
          "createdAt"
        ],
        "title": "OrderInDB"
      },
...
      "OrderItemInDB": {
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid",
            "title": "Productid"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "priceSnapshot": {
            "type": "string",
            "title": "Pricesnapshot"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "imageUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imageurl"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "orderId": {
            "type": "string",
            "format": "uuid",
            "title": "Orderid"
          }
        },
        "type": "object",
        "required": [
          "productId",
          "quantity",
          "priceSnapshot",
          "name",
          "id",
          "orderId"
        ],
        "title": "OrderItemInDB"
      },
...
      "ProductCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Price"
          },
          "discount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discount"
          },
          "characteristics": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Characteristics"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "category_id": {
            "type": "string",
            "format": "uuid",
            "title": "Category Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "slug",
          "price",
          "category_id"
        ],
        "title": "ProductCreate"
      },
      "ProductInDB": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "price": {
            "type": "string",
            "title": "Price"
          },
          "discount": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discount"
          },
          "characteristics": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Characteristics"
          },
          "imageUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imageurl"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "title": "Categoryid"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CategoryInDB"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updatedat"
          },
          "timesOrdered": {
            "type": "integer",
            "title": "Timesordered"
          }
        },
        "type": "object",
        "required": [
          "name",
          "slug",
          "price",
          "categoryId",
          "id",
          "createdAt",
          "timesOrdered"
        ],
        "title": "ProductInDB"
      },
      "ProductUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "discount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discount"
          },
          "characteristics": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Characteristics"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "categoryId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Categoryid"
          }
        },
        "type": "object",
        "title": "ProductUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Development server"
    }
  ]
}